requirements.txt
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.0
pandas
numpy
scikit-learn
tqdm
Pillow
